requirements.txt
torch>=2.0.0
ultralytics>=8.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
transformers>=4.35.0
streamlit>=1.28.0
pillow>=10.0.0
numpy>=1.24.0
hf_xet
qwen-vl-utils
qwen-vl-utils[decord]==0.0.8
accelerate